Move met_eireann coordinator to separate module
